Move integration tests to their own folder
